OMFG I can't believe it isn't borked!
Short version: basically, the app fulfills very nearly the bear minimum
now. I need to do some state preservation on the palette after all, but
once that is functioning, an ~C grade project seems to be done.

Other things I need to fix:
1) Swap base colors to CMYK? I seem to be doing subtractive colors
somehow, unintentionally, but the base colors need to be something nicer
so you can get all the shades.
2) Add hand tool, and possibly fine tune the drawing (since things seem to
stray off screen a bit - offset probably needs to be reset is what it is!)
3) Remove colors?