Presently the LibraryListView is somewhat setup. It's displaying roughly all the expected information, and it can get more images from the camera. Looks like shit, and has a good number of bugs, but looks like it will work overall.
Also, refactoring to the Library in the model: turns out, opening a file
every time you want to get a thumbnail and everything is FUCKING SLOW. Go
figure? So, now all that data is being cached in the library instead.
Shouldn't be too bad, although it is unfortunate I have to eat up yet more
memory with this fucking class.