Refactored code (yet again) to fix up the way we communicate that we want to search over collections rather than all cards. Code is now in params (though apparently that was already there in a previous version? Didn't realize, so I just started using it).
Anyway, with that in place, I've got these meaningful updates:

1) Advanced Search pages look substantially better, esp. in landscape mode
(apparently Android just magically takes care of figuring out orientation
for you if you do it properly!). Still need some work for tuning, but good
enough for a dry run, I'd say. Or close, anyway.

2) You can now search over all cards or over collections. Includes the
option to search over ALL collections, or just one collection at a time.
Presently the AdvancedSearchActivity has been separated from the
CollectionSearchActivity, which is not the nicest (since it means repeated
XML layout code at the very least), but it's substantially cleaner and
easier to maintain IMHO than reusing an activity or multiple purposes.
Probably still the best way to go, even if somewhat frustrating in terms
of code maintainence.

3) Viewing a specific collection works! I don't recall if that was in the
previous commit or not, but I know that I retooled it a litte bit this
morning, so I'll just document it here.

This is actually just about the entire control flow, just need to provide
a way to add cards to collections, then start cleaning up the
CardDetailFragment to be nicer and display better info. Generally
speaking, though, the project is very nearly complete. :D